pyproject.toml
[project]
name = "partyDecider"
version = "0.1"
description = "Decider for party games"
requires-python = ">=3.10"
authors = [
    {name = "Jyrki Launonen"},
]

dependencies = [
    "python-dotenv~=1.0.0",
    "fastapi~=0.95.0",
    "uvicorn~=0.21.0",
]

[project.optional-dependencies]
dev = [
    "black~=23.3.0",
    "isort~=5.12.0",
    "pip-tools~=6.13.0",
]
# Required by build.py on python 3.10
py310 = [
    "tomlkit",
]

[project.scripts]
main = "backend.__main__:main"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.setuptools]
py-modules = ["backend"]
